+++ eliis-front-end/public/index.html
@@ -1,27 +1,49 @@
 <!DOCTYPE html>
 <html lang="en" class="sr" dir="ltr">
   <head>
-    <meta charset="utf-8" />
-    <link rel="icon" type="image/svg+xml" href="https://upload.wikimedia.org/wikipedia/commons/0/0b/Eo_circle_light-green_white_letter-e.svg" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <meta name="theme-color" content="#000000" />
+    <meta charset="utf-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <style>
+      /* Inline critical CSS goes here */
+      body {
+        margin: 0;
+        font-family: Arial, sans-serif;
+      }
+    </style>
+    <link rel="canonical" href="https://vp-web-development-showcase.netlify.app/">
+    <link rel="icon" type="image/svg+xml" href="https://upload.wikimedia.org/wikipedia/commons/0/0b/Eo_circle_light-green_white_letter-e.svg">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <meta name="theme-color" content="#000000">
     <meta
       name="description"
       content="ELIIS in an online information system for pre-school organizations."
-    />
-    <!--Apple Web App Meta-->
+    >
     <meta name="apple-mobile-web-app-capable" content="yes">
     <meta name="apple-mobile-web-app-status-bar-style" content="black">
     <meta name="format-detection" content="telephone=yes">
-    <meta name = "viewport" content = "width = 320, initial-scale = 2.3, user-scalable = no">
-    <!--Windows tiles-->
+    <meta name="viewport" content="width=320, initial-scale=2.3">
+
     <meta name="msapplication-TileImage" content="./pic.png">
     <meta name="msapplication-TileColor" content="#ffffff">
 
+    <meta property="og:title" content="ELIIS Calendar">
+    <meta property="og:description" content="ELIIS Calendar is a feature-rich calendar application designed to streamline your scheduling and event management needs. Whether you're an individual managing personal events or a team coordinating professional schedules, ELIIS Calendar provides a user-friendly and customizable platform.">
+    <meta property="og:image" content="./pic.png">
+    <meta property="og:url" content="https://vp-web-development-showcase.netlify.app/">
+    <meta property="og:type" content="website">
+    
+    <meta name="twitter:card" content="Summary">
+    <meta name="twitter:title" content="ELIIS Calendar">
+    <meta name="twitter:description" content="ELIIS Calendar is a feature-rich calendar application designed to streamline your scheduling and event management needs. Whether you're an individual managing personal events or a team coordinating professional schedules, ELIIS Calendar provides a user-friendly and customizable platform.">
+    <meta name="twitter:image" content="./pic.png">
+    <meta name="twitter:url" content="https://vp-web-development-showcase.netlify.app/">
     <title>Eliis calendar</title>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
-    <div id="root"></div>
+    <main>
+      <div id="root"></div>
+    </main>
   </body>
 </html>
