+++ eliis-front-end/src/components/pages/Application Components/Applications/application.css
