+++ eliis-front-end/src/components/pages/Events/Calendar/Calendar.js
@@ -5,14 +5,20 @@ import FullCalendar from '@fullcalendar/react';
 import dayGridPlugin from '@fullcalendar/daygrid';
 import timeGridPlugin from '@fullcalendar/timegrid';
 import interactionPlugin from '@fullcalendar/interaction';
-import { INITIAL_EVENTS, createEventId } from './event-utils';
+import Button from 'react-bootstrap/Button';
+import Modal from 'react-bootstrap/Modal';
+import { Dropdown } from 'react-bootstrap';
 import './CalendarS.css';
 import './index.css';
-import StaticExample from './modal';
+import Icon from '@mdi/react';
+import { mdiCheck } from '@mdi/js';
 
 const SERVER_URL = 'http://localhost:3001';
 
 export default function DemoApp() {
+  let eventGuid = 0
+  let todayStr = new Date().toISOString().replace(/T.*$/, '') // YYYY-MM-DD of today
+
   const [weekendsVisible, setWeekendsVisible] = useState(true);
   const [currentEvents, setCurrentEvents] = useState([]);
   const [calendarRef, setCalendarRef] = useState(null);
@@ -25,6 +31,89 @@ export default function DemoApp() {
   const handleCloseModal = () => {
     setIsModalOpen(false);
   };
+
+  const typeMenu = [
+    { color: 'Coral', label: 'Education project' },
+    { color: 'Olive-Green', label: 'Theatre/Concert' },
+    { color: 'Rose-Pink', label: 'Meeting' },
+    { color: 'Tomato-Red', label: 'Training' },
+    { color: 'Muted-Pink', label: 'Joint event' },
+    { color: 'Mustard-Yellow', label: 'Class event' },
+    { color: 'Teal', label: 'Learning activity' },
+    { color: 'Sky-Blue', label: 'Other event' },
+    { color: 'White-Red-Outline', label: 'Public holiday' },
+  ];
+  
+  const EventCalendarHeader = () => {
+    return (
+      <div className="flex flex-col gap-2">
+        <h1 className='text-3xl'>Event Calendar</h1>
+        <h3>Types</h3>
+          {typeMenu.map((item, index) => (
+            <div key={index} className="flex items-center gap-2">
+              <span className={`rounded-full p-1 flex items-center justify-center font-bold w-6 h-6 text-xs bg-${item.color} text-${index === typeMenu.length - 1 ? 'black' : 'white'}`}>
+                <Icon 
+                  path={mdiCheck} 
+                  size={1} 
+                  color="currentColor"
+                />
+              </span>
+              <span data-testid={item.label}>{item.label}</span>
+            </div>
+          ))}
+      </div>
+    );
+  }
+
+  const INITIAL_EVENTS = [
+    {
+      id: createEventId(),
+      title: 'All-day event',
+      start: todayStr
+    },
+    {
+      id: createEventId(),
+      title: 'Timed event',
+      start: todayStr + 'T12:00:00'
+    }
+  ]
+  
+  function createEventId() {
+    return String(eventGuid++)
+  }
+
+  function StaticExample({ onClose }) {
+    return (
+      <div className="modal show" style={{ display: 'block', position: 'initial' }}>
+        <Modal.Dialog>
+          <Modal.Header closeButton onClick={onClose}>
+            <Modal.Title>Create Event</Modal.Title>
+          </Modal.Header>
+  
+          <Modal.Body>
+            <div>
+              <Dropdown>
+                <Dropdown.Toggle variant="success" id="dropdown-basic">
+                  Choose type of event
+                </Dropdown.Toggle>
+  
+                <Dropdown.Menu>
+                  <Dropdown.Item>First</Dropdown.Item>
+                </Dropdown.Menu>
+              </Dropdown>
+            </div>
+          </Modal.Body>
+  
+          <Modal.Footer>
+            <Button variant="secondary" onClick={onClose}>
+              Close
+            </Button>
+            <Button variant="primary">Save changes</Button>
+          </Modal.Footer>
+        </Modal.Dialog>
+      </div>
+    );
+  }
 
   const handleDateSelect = (selectInfo) => {
     let title = prompt('Please enter a new title for your event');
@@ -102,7 +191,6 @@ export default function DemoApp() {
           selectable={true}
           selectMirror={true}
           dayMaxEvents={true}
-          updateSize={true}
           height={'auto'}
           weekends={weekendsVisible}
           initialEvents={INITIAL_EVENTS}
