+++ eliis-front-end/src/components/pages/Events/Calendar/CalendarS.css
@@ -23,3 +23,93 @@
   color: black;
   text-decoration: none;
 }
+
+/* Common styles */
+.flex {
+  display: flex;
+}
+
+.flex-col {
+  flex-direction: column;
+}
+
+.items-center {
+  align-items: center;
+}
+
+.text-3xl {
+  font-size: 1.875rem;
+  line-height: 2.25rem;
+}
+
+.font-bold {
+  font-weight: 700;
+}
+
+.text-xs {
+  font-size: 0.75rem;
+  line-height: 1rem;
+}
+
+.rounded-full {
+  border-radius: 9999px;
+}
+
+.justify-center {
+  justify-content: center;
+}
+
+.w-6 {
+  width: 1.5rem;
+}
+
+.h-6 {
+  height: 1.5rem;
+}
+
+/* Background colors */
+.bg-Coral {
+  --tw-bg-opacity: 1;
+  background-color: rgb(247 158 128 / var(--tw-bg-opacity));
+}
+
+.bg-Olive-Green {
+  background-color: rgb(166, 194, 100);
+}
+
+.bg-Rose-Pink {
+  --tw-bg-opacity: 1;
+  background-color: rgb(224 130 177 / var(--tw-bg-opacity));
+}
+
+.bg-Tomato-Red {
+  --tw-bg-opacity: 1;
+  background-color: rgb(235 94 93 / var(--tw-bg-opacity));
+}
+
+.bg-Muted-Pink {
+  --tw-bg-opacity: 1;
+  background-color: rgb(204 121 121 / var(--tw-bg-opacity));
+}
+
+.bg-Mustard-Yellow {
+  --tw-bg-opacity: 1;
+  background-color: rgb(239 173 78 / var(--tw-bg-opacity));
+}
+
+.bg-Teal {
+  --tw-bg-opacity: 1;
+  background-color: rgb(115 202 194 / var(--tw-bg-opacity));
+}
+
+.bg-Sky-Blue {
+  --tw-bg-opacity: 1;
+  background-color: rgb(92 192 222 / var(--tw-bg-opacity));
+}
+
+/* Special style for white background with red border */
+.bg-White-Red-Outline {
+  --tw-bg-opacity: 1;
+  background-color: rgb(255, 255, 255, var(--tw-bg-opacity));
+  border: 1px solid red;
+}
