--- eliis-front-end/src/components/pages/Events/Calendar/modal.js
@@ -1,39 +1,0 @@
-import React from 'react';
-import Button from 'react-bootstrap/Button';
-import Modal from 'react-bootstrap/Modal';
-import { Dropdown } from 'react-bootstrap';
-
-function StaticExample({ onClose }) {
-  return (
-    <div className="modal show" style={{ display: 'block', position: 'initial' }}>
-      <Modal.Dialog>
-        <Modal.Header closeButton onClick={onClose}>
-          <Modal.Title>Create Event</Modal.Title>
-        </Modal.Header>
-
-        <Modal.Body>
-          <div>
-            <Dropdown>
-              <Dropdown.Toggle variant="success" id="dropdown-basic">
-                Choose type of event
-              </Dropdown.Toggle>
-
-              <Dropdown.Menu>
-                <Dropdown.Item>First</Dropdown.Item>
-              </Dropdown.Menu>
-            </Dropdown>
-          </div>
-        </Modal.Body>
-
-        <Modal.Footer>
-          <Button variant="secondary" onClick={onClose}>
-            Close
-          </Button>
-          <Button variant="primary">Save changes</Button>
-        </Modal.Footer>
-      </Modal.Dialog>
-    </div>
-  );
-}
-
-export default StaticExample;
